requirements.txt
alembic==1.7.5
azure-common==1.1.28
azure-core==1.22.0
azure-search-documents==11.2.1
barnum==0.5.1
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
cryptography==36.0.1
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-User==1.0.2.2
Flask-WTF==1.0.0
future==0.18.2
idna==3.3
isodate==0.6.1
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
msrest==0.6.21
mysql-connector-python==8.0.27
oauthlib==3.2.0
passlib==1.7.4
protobuf==3.19.3
pycparser==2.21
requests==2.27.1
requests-oauthlib==1.3.1
schedule==1.1.0
six==1.16.0
SQLAlchemy==1.4.29
termcolor==1.1.0
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
WTForms==3.0.1
